+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes
 
 0.21.3 (unreleased)
 
+ - Fix `ssdp_listener.SsdpDeviceTracker` to update device's headers upon ssdp:byebye advertisement (@chishm)
 
 0.21.2 (2021-09-12)